requirements.txt
bokeh
ghp-import
holoviews
jupyter-bokeh
jupyter-book
lxml[html_clean]
markdown 
markdown-it-py           
matplotlib-inline
myst-nb                  
myst-parser              
nbclient                 
nbconvert                
nbformat                 
nbsphinx
nbsphinx                 
ncurses                  
nest-asyncio             
numpy                    
pandas                   
pandocfilters            
panel                    
param                    
parso                    
pexpect                  
pillow                   
playwright               
plotly                   
prompt-toolkit           
psutil                   
ptyprocess               
pyaml                    
pybtex                   
pybtex-docutils          
pydata-sphinx-theme      
pytz                     
pyyaml                   
requests                 
rich                     
rpds-py                  
scipy                    
sphinx                   
sphinx_proof
sphinx-book-theme        
sphinx-comments          
sphinx-copybutton
sphinx-design
sphinx-design            
sphinx-external-toc      
sphinx-jupyterbook-latex 
sphinx-multitoc-numbering
sphinx-prolog
sphinx-proof             
sphinx-pyscript          
sphinx-thebe             
sphinx-togglebutton
sphinx-togglebutton      
sphinxcontrib-applehelp  
sphinxcontrib-bibtex     
sphinxcontrib-devhelp    
sphinxcontrib-htmlhelp   
sphinxcontrib-jsmath     
sphinxcontrib-mermaid    
sphinxcontrib-qthelp     
sphinxcontrib.mermaid
sqlalchemy               
tqdm                     
typer                    